Handling large numbers in ex01

+++ ex01/RPN.cpp
@@ -6,14 +6,11 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/11 15:29:42 by rnovotny          #+#    #+#             */
-/*   Updated: 2026/03/11 16:35:08 by rnovotny         ###   ########.fr       */
+/*   Updated: 2026/05/06 19:36:50 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "RPN.hpp"
-#include <sstream>
-#include <stdexcept>
-#include <cstdlib>
 
 RPN::RPN() {}
 
@@ -29,17 +26,17 @@ RPN::RPN(const std::string& expression)
 			if (_stack.size() < 2)
 				throw std::runtime_error("Invalid expression");
 			
-			int b = _stack.top();
+			long b = _stack.top();
 			_stack.pop();
-			int a = _stack.top();
+			long a = _stack.top();
 			_stack.pop();
 			
-			int result = performOperation(a, b, token[0]);
+			long result = performOperation(a, b, token[0]);
 			_stack.push(result);
 		}
 		else if (isNumber(token))
 		{
-			int num = std::atoi(token.c_str());
+			long num = std::atol(token.c_str());
 			if (num >= 10 || num < 0)
 				throw std::runtime_error("Number must be < 10");
 			_stack.push(num);
@@ -62,7 +59,7 @@ RPN& RPN::operator=(const RPN& other)
 
 RPN::~RPN() {}
 
-int RPN::calculate()
+long RPN::calculate()
 {
 	if (_stack.size() != 1)
 		throw std::runtime_error("Invalid expression");
@@ -88,7 +85,7 @@ bool RPN::isNumber(const std::string& token) const
 	return true;
 }
 
-int RPN::performOperation(int a, int b, char op)
+long RPN::performOperation(long a, long b, char op)
 {
 	switch (op)
 	{
