+++ ex01/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/11 15:29:27 by rnovotny          #+#    #+#             */
-/*   Updated: 2026/03/11 16:35:08 by rnovotny         ###   ########.fr       */
+/*   Updated: 2026/05/06 19:36:55 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,14 @@
 
 # include <stack>
 # include <string>
+# include <sstream>
+# include <stdexcept>
+# include <cstdlib>
 
 class RPN
 {
 private:
-	std::stack<int> _stack;
+	std::stack<long> _stack;
 	
 	RPN();
 	
@@ -29,12 +32,12 @@ public:
 	RPN& operator=(const RPN& other);
 	~RPN();
 	
-	int calculate();
+	long calculate();
 	
 private:
 	bool isOperator(char c) const;
 	bool isNumber(const std::string& token) const;
-	int performOperation(int a, int b, char op);
+	long performOperation(long a, long b, char op);
 };
 
 #endif
